requirement.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
astor==0.8.1
attrdict==2.0.1
attrs==23.2.0
Babel==2.15.0
bce-python-sdk==0.9.14
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cssselect==1.2.0
cssutils==2.11.1
cycler==0.12.1
Cython==3.0.10
dashscope==1.19.2
dataclasses-json==0.6.7
decorator==5.1.1
distro==1.9.0
dnspython==2.6.1
easyocr==1.7.1
elastic-transport==8.13.1
elasticsearch==8.14.0
email_validator==2.1.1
environs==9.5.0
et-xmlfile==1.1.0
fastapi==0.111.0
fastapi-cli==0.0.4
filelock==3.14.0
fire==0.6.0
Flask==3.0.3
flask-babel==4.0.0
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.6.0
future==1.0.0
grpcio==1.63.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.6
imageio==2.34.1
imgaug==0.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==2.4
kiwisolver==1.4.5
langchain==0.2.12
langchain-community==0.2.11
langchain-core==0.2.28
langchain-text-splitters==0.2.1
langsmith==0.1.98
lazy_loader==0.4
lmdb==1.4.1
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
matplotlib==3.9.0
mdurl==0.1.2
milvus-lite==2.4.6
more-itertools==10.3.0
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.3
ninja==1.11.1.1
numpy==1.26.4
openai==1.35.10
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opencv-python-headless==4.10.0.82
openpyxl==3.1.3
opt-einsum==3.3.0
orjson==3.10.3
packaging==23.2
paddleocr==2.7.3
paddlepaddle==2.6.1
pandas==2.2.2
pdf2docx==0.5.8
pillow==10.3.0
premailer==3.10.0
protobuf==5.27.0
psutil==5.9.8
pyclipper==1.3.0.post5
pycryptodome==3.20.0
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.18.0
PyJWT==2.8.0
pymilvus==2.4.3
PyMuPDF==1.24.5
PyMuPDFb==1.24.3
PyMySQL==1.1.1
pyparsing==3.1.2
pypdf==4.3.1
python-bidi==0.4.2
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
rapidfuzz==3.9.3
rarfile==4.2
requests==2.31.0
rich==13.7.1
scikit-image==0.23.2
scipy==1.13.1
setuptools==70.0.0
shapely==2.0.4
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.30
starlette==0.37.2
sympy==1.12.1
tenacity==8.3.0
termcolor==2.4.0
tifffile==2024.5.22
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.4
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.1
uvicorn==0.30.1
uvloop==0.19.0
visualdl==2.5.3
volcengine-python-sdk==1.0.90
watchfiles==0.22.0
websockets==12.0
Werkzeug==3.0.3
xinference-client==0.13.0
yarl==1.9.4
zhipuai==2.0.1
